setup journey entry page
