+++ app-web/journeys/getting-started-as-a-developer.md
@@ -1,0 +1,17 @@
+---
+id: Beginner Guide to Developing on the Platform
+resourceType: Journey
+description: This guide will help onboard you onto the platform.
+tags: 
+- openshift
+- guide
+- journey
+- noob
+- newbie
+- new developer
+---
+
+You and more correctly, your team has or is just about to be on-boarded onto the Pathfinder platform and need to get started. One of the questions that is often asked is _"How do I get Started?"_
+
+This guide aims to be help you and your team in this journey. It highlights several concepts, pieces of tech, as well as helpful tools & resources. It does not aim to be an exhaustive guide initially but it does intend to be one that is consistently improved to help you, the Developer, succeed. 
+
