+++ app-web/src/utils/helpers.js
@@ -374,3 +374,14 @@ export const isLocalHost = () => {
 
   return origin.indexOf('localhost') >= 0 || origin.indexOf('0.0.0.0') >= 0;
 };
+
+/**
+ * reduces the connects with node field to subway stops usable by the subwayline component
+ * @param {Array} connections the subway stops found from registryJourneyJson.connectsWith
+ */
+export const reduceJourneyToSubwayLine = connections =>
+  connections.map((connection, index) => ({
+    name: connection.fields.title,
+    to: connection.path,
+    variant: index % 2 === 0 ? 'up' : 'down',
+  }));
